+++ Solution/Utils/sync/Semaphores.cpp
@@ -6,12 +6,12 @@
 #include "../device/LogicalDevice.h"
 
 namespace vk2 {
-Semaphores::Semaphores(const uint32_t semaphoreCount) {
+Semaphores::Semaphores(const size_t semaphoreCount) {
     assert(semaphoreCount > 0);
     
     vk::Device device(LogicalDevice::device());
     mSemaphores.reserve(semaphoreCount);
-    for (uint32_t i = 0; i < semaphoreCount; ++i) {
+    for (size_t i = 0; i < semaphoreCount; ++i) {
         mSemaphores.emplace_back(device.createSemaphoreUnique({}));
     }
 }
